Add files via upload
Creating a simple two-page website to encourage your loved ones to pursue their goals can be a thoughtful and personalized way to provide support. there are two ways to add gif 
Offline Version:

"For the offline version, include the provided GIFs to enhance the user experience. Utilize them to visually represent key messages and make the content more engaging."
Online Version:

"In the online version, include a link within the HTML code. Ensure the link directs users to relevant online resources or additional information that complements the content on the website."